+++ NEWS.md
@@ -1,4 +1,20 @@
-# wbpip 0.0.3.9000
+# wbpip (development version)
+
+* Add new functions to compute the following indicators:  
+  - Number of poor
+  - Average shortfall
+  - Total shortfall
+  - Income gap ratio
+  - Palma ratio
+  - Top 10 / Bottom 10 ratio
+
+# wbpip 0.1.0
+
+## New features
+* Updated SPL function
+* Small optimization
+
+# wbpip 0.0.4
 
 ## New features
 
@@ -11,6 +27,8 @@
 * create a separate function (`weighted_average_poverty_stats`) for repetitive calculation in adjust poverty stats and add corresponding test case for it. 
 
 ## Bug fixes
+* Fix duplicated values being returned for group data when poverty line when no
+Lorenz fit was successful
 
 # wbpip 0.0.3